@@ -16,30 +16,44 @@
         }
     },
     "require": {
-        "keboola-legacy/php-encryption": "^0.2",
-        "keboola/php-encryption": "^0.2",
-        "defuse/php-encryption": "^2.3",
-        "php": ">=7",
+        "php": ">=7.4",
+        "ext-zlib": "*",
+        "ext-json": "*",
         "aws/aws-sdk-php": "^3.209",
-        "vkartaviy/retry": "^0.2.2",
-        "phpseclib/mcrypt_compat": "^1.0",
+        "defuse/php-encryption": "^2.3",
         "keboola/azure-key-vault-client": "^2.0",
-        "ext-zlib": "*"
+        "keboola/common-exceptions": "^1.2",
+        "phpseclib/mcrypt_compat": "^1.0",
+        "vkartaviy/retry": "^0.2.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7|^6.5|^7.3",
-        "php-parallel-lint/php-parallel-lint": "^1.3"
+        "infection/infection": "^0.18.2",
+        "keboola/coding-standard": "^13.0",
+        "phpstan/phpstan": "^1.8",
+        "phpstan/phpstan-phpunit": "^1.0.0",
+        "phpunit/phpunit": "^9.5"
     },
     "scripts": {
-        "tests": "phpunit --coverage-clover build/logs/clover.xml --coverage-xml=build/logs/coverage-xml --log-junit=build/logs/phpunit.junit.xml",
-        "phplint": "parallel-lint -j 10 --exclude vendor .",
+        "tests": "phpunit --coverage-clover /tmp/build-logs/clover.xml --coverage-xml=/tmp/build-logs/coverage-xml --log-junit=/tmp/build-logs/phpunit.junit.xml",
+        "phpcs": "phpcs --extensions=php src tests",
+        "phpstan": "phpstan analyse --no-progress -c phpstan.neon",
+        "infection": "infection --threads=4 --min-covered-msi=90 --coverage=/tmp/build-logs --only-covered",
         "build": [
-            "@phplint",
-            "@tests"
+            "@tests",
+            "@infection"
         ],
         "ci": [
             "@composer validate --no-check-all --strict",
             "@build"
         ]
+    },
+    "config": {
+        "sort-packages": true,
+        "process-timeout": 3600,
+        "lock": false,
+        "allow-plugins": {
+            "infection/extension-installer": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
     }
 }
