@@ -2,10 +2,9 @@
 
 namespace Keboola\ObjectEncryptor;
 
+use Exception;
 use Keboola\ObjectEncryptor\Exception\ApplicationException;
 use Keboola\ObjectEncryptor\Exception\UserException;
-use Keboola\ObjectEncryptor\Legacy\Encryptor;
-use Keboola\ObjectEncryptor\Legacy\Wrapper\BaseWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ComponentAKVWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ComponentKMSWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ConfigurationAKVWrapper;
@@ -13,6 +12,8 @@ use Keboola\ObjectEncryptor\Wrapper\ConfigurationKMSWrapper;
 use Keboola\ObjectEncryptor\Wrapper\CryptoWrapperInterface;
 use Keboola\ObjectEncryptor\Wrapper\ProjectAKVWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ProjectKMSWrapper;
+use stdClass;
+use Throwable;
 
 class ObjectEncryptor
 {
@@ -20,33 +21,16 @@ class ObjectEncryptor
      * List of known wrappers.
      * @var CryptoWrapperInterface[]
      */
-    private $wrappers = [];
+    private array $wrappers = [];
 
     /**
-     * Encryptor used only for decrypting legacy cipher texts.
-     * @var Encryptor|null
-     */
-    private $legacyEncryptor = null;
-
-
-    /**
-     * ObjectEncryptor constructor.
-     * @param Encryptor|null $legacyEncryptor Optional legacy decryptor.
-     */
-    public function __construct(Encryptor $legacyEncryptor = null)
-    {
-        $this->legacyEncryptor = $legacyEncryptor;
-    }
-
-    /**
-     * @param string|array|\stdClass $data Data to encrypt
-     * @param string $wrapperName Class name of encryptor wrapper
-     * @return mixed
+     * @param string|array|stdClass $data Data to encrypt
+     * @return string|array|stdClass
      * @throws ApplicationException
+     * @throws UserException
      */
-    public function encrypt($data, $wrapperName = BaseWrapper::class)
+    public function encrypt($data, string $wrapperName)
     {
-        /** @var BaseWrapper $wrapper */
         foreach ($this->wrappers as $cryptoWrapper) {
             if (get_class($cryptoWrapper) == $wrapperName) {
                 $wrapper = $cryptoWrapper;
@@ -62,31 +46,27 @@ class ObjectEncryptor
         if (is_array($data)) {
             return $this->encryptArray($data, $wrapper);
         }
-        if (is_object($data) && get_class($data) == \stdClass::class) {
+        if (is_object($data) && get_class($data) == stdClass::class) {
             return $this->encryptObject($data, $wrapper);
         }
         throw new ApplicationException('Only stdClass, array and string are supported types for encryption.');
     }
 
     /**
-     * @param mixed $data
-     * @return mixed
+     * @param string|array|stdClass $data
+     * @return string|array|stdClass
      * @throws ApplicationException
      * @throws UserException
      */
     public function decrypt($data)
     {
         if (is_scalar($data)) {
-            try {
-                return $this->decryptValue($data);
-            } catch (\InvalidCiphertextException $e) {
-                throw new UserException($e->getMessage(), $e);
-            }
+            return $this->decryptValue($data);
         }
         if (is_array($data)) {
             return $this->decryptArray($data);
         }
-        if (is_a($data, \stdClass::class) && (get_class($data) == \stdClass::class)) {
+        if (is_a($data, stdClass::class) && (get_class($data) == stdClass::class)) {
             return $this->decryptObject($data);
         }
         throw new ApplicationException('Only stdClass, array and string are supported types for decryption.');
@@ -106,7 +86,7 @@ class ObjectEncryptor
         $this->wrappers[$wrapper->getPrefix()] = $wrapper;
     }
 
-    public function getRegisteredComponentWrapperClass()
+    public function getRegisteredComponentWrapperClass(): string
     {
         foreach ($this->wrappers as $wrapper) {
             if (get_class($wrapper) === ComponentKMSWrapper::class || get_class($wrapper) === ComponentAKVWrapper::class) {
@@ -116,7 +96,7 @@ class ObjectEncryptor
         throw new ApplicationException('No Component wrappers registered.');
     }
 
-    public function getRegisteredProjectWrapperClass()
+    public function getRegisteredProjectWrapperClass(): string
     {
         foreach ($this->wrappers as $wrapper) {
             if (get_class($wrapper) === ProjectKMSWrapper::class || get_class($wrapper) === ProjectAKVWrapper::class) {
@@ -126,7 +106,7 @@ class ObjectEncryptor
         throw new ApplicationException('No Project wrappers registered.');
     }
 
-    public function getRegisteredConfigurationWrapperClass()
+    public function getRegisteredConfigurationWrapperClass(): string
     {
         foreach ($this->wrappers as $wrapper) {
             if (get_class($wrapper) === ConfigurationKMSWrapper::class || get_class($wrapper) === ConfigurationAKVWrapper::class) {
@@ -138,16 +118,10 @@ class ObjectEncryptor
 
     /**
      * Find a wrapper to decrypt a given cipher.
-     * @param string $value Cipher text
-     * @return CryptoWrapperInterface|null
-     * @throws ApplicationException
      */
-    private function findWrapper($value)
+    private function findWrapper(string $value): ?CryptoWrapperInterface
     {
         $selectedWrapper = null;
-        if (empty($this->wrappers)) {
-            throw new ApplicationException('There are no wrappers registered for the encryptor.');
-        }
         foreach ($this->wrappers as $wrapper) {
             if (substr($value, 0, mb_strlen($wrapper->getPrefix())) == $wrapper->getPrefix()) {
                 $selectedWrapper = $wrapper;
@@ -156,56 +130,28 @@ class ObjectEncryptor
         return $selectedWrapper;
     }
 
-    private function decryptLegacy($value)
-    {
-        /* @ is intentional to suppress warnings from invalid cipher texts which
-        are handled by checking return === false */
-        if ($this->legacyEncryptor) {
-            $ret = @$this->legacyEncryptor->decrypt($value);
-            if ($ret === false) {
-                throw new UserException('Value is not an encrypted value.');
-            } else {
-                return $ret;
-            }
-        } else {
-            throw new UserException('Value is not an encrypted value.');
-        }
-    }
-
-    /**
-     * @param string $value Value to decrypt.
-     * @return string Decrypted value.
-     * @throws ApplicationException
-     * @throws UserException
-     */
-    private function decryptValue($value)
+    private function decryptValue(string $value): string
     {
         $wrapper = $this->findWrapper($value);
         if (!$wrapper) {
-            return $this->decryptLegacy($value);
+            throw new UserException(sprintf('Value "%s" is not an encrypted value.', $value), 0);
         } else {
             try {
                 return $wrapper->decrypt(substr($value, mb_strlen($wrapper->getPrefix())));
-            } catch (\InvalidCiphertextException $e) {
-                // this is for legacy wrappers
-                throw new UserException("Value $value is not an encrypted value.", $e);
             } catch (UserException $e) {
-                throw new UserException("Value $value is not an encrypted value.", $e);
-            } catch (\Exception $e) {
+                throw new UserException(sprintf('Value "%s" is not an encrypted value.', $value), $e->getCode(), $e);
+            } catch (Throwable $e) {
                 // decryption failed for more serious reasons
-                throw new ApplicationException('Decryption failed: ' . $e->getMessage(), $e);
+                throw new ApplicationException('Decryption failed: ' . $e->getMessage(), $e->getCode(), $e);
             }
         }
     }
 
     /**
-     * @param string $key Array key
-     * @param array|string|\stdClass|null $value Value to encrypt.
-     * @param CryptoWrapperInterface $wrapper
-     * @return array|string|\stdClass|null
-     * @throws ApplicationException
+     * @param array|string|stdClass|null $value Value to encrypt.
+     * @return array|string|stdClass|null
      */
-    private function encryptItem($key, $value, CryptoWrapperInterface $wrapper)
+    private function encryptItem(string $key, $value, CryptoWrapperInterface $wrapper)
     {
         if (is_scalar($value) || is_null($value)) {
             if (substr($key, 0, 1) == '#') {
@@ -215,7 +161,7 @@ class ObjectEncryptor
             }
         } elseif (is_array($value)) {
             return $this->encryptArray($value, $wrapper);
-        } elseif (is_object($value) && get_class($value) == \stdClass::class) {
+        } elseif (is_object($value) && get_class($value) == stdClass::class) {
             return $this->encryptObject($value, $wrapper);
         } else {
             throw new ApplicationException(
@@ -224,13 +170,7 @@ class ObjectEncryptor
         }
     }
 
-    /**
-     * @param string $value Value to encrypt.
-     * @param CryptoWrapperInterface $wrapper Ciphering wrapper.
-     * @return string Encrypted value.
-     * @throws ApplicationException
-     */
-    private function encryptValue($value, CryptoWrapperInterface $wrapper)
+    private function encryptValue(string $value, CryptoWrapperInterface $wrapper): string
     {
         // return self if already encrypted with any wrapper
         if ($this->findWrapper($value)) {
@@ -239,18 +179,12 @@ class ObjectEncryptor
 
         try {
             return $wrapper->getPrefix() . $wrapper->encrypt($value);
-        } catch (\Exception $e) {
-            throw new ApplicationException('Encryption failed: ' . $e->getMessage(), $e);
+        } catch (Throwable $e) {
+            throw new ApplicationException('Encryption failed: ' . $e->getMessage(), $e->getCode(), $e);
         }
     }
 
-    /**
-     * @param array $data
-     * @param CryptoWrapperInterface $wrapper
-     * @return array
-     * @throws ApplicationException
-     */
-    private function encryptArray(array $data, CryptoWrapperInterface $wrapper)
+    private function encryptArray(array $data, CryptoWrapperInterface $wrapper): array
     {
         $result = [];
         foreach ($data as $key => $value) {
@@ -259,15 +193,9 @@ class ObjectEncryptor
         return $result;
     }
 
-    /**
-     * @param \stdClass $data
-     * @param CryptoWrapperInterface $wrapper
-     * @return \stdClass
-     * @throws ApplicationException
-     */
-    private function encryptObject(\stdClass $data, CryptoWrapperInterface $wrapper)
+    private function encryptObject(stdClass $data, CryptoWrapperInterface $wrapper): stdClass
     {
-        $result = new \stdClass();
+        $result = new stdClass();
         foreach (get_object_vars($data) as $key => $value) {
             $result->{$key} = $this->encryptItem($key, $value, $wrapper);
         }
@@ -275,27 +203,24 @@ class ObjectEncryptor
     }
 
     /**
-     * @param string $key Key in array.
-     * @param array|string|\stdClass|null $value Value to decrypt.
-     * @return array|string|\stdClass|null Decrypted value.
-     * @throws ApplicationException
-     * @throws UserException
+     * @param array|string|stdClass|null $value Value to decrypt.
+     * @return array|string|stdClass|null Decrypted value.
      */
-    private function decryptItem($key, $value)
+    private function decryptItem(string $key, $value)
     {
         if (is_scalar($value) || is_null($value)) {
             if (substr($key, 0, 1) == '#') {
                 try {
                     return $this->decryptValue((string)$value);
                 } catch (UserException $e) {
-                    throw new UserException("Invalid cipher text for key $key " . $e->getMessage(), $e);
+                    throw new UserException("Invalid cipher text for key $key " . $e->getMessage(), $e->getCode(), $e);
                 }
             } else {
                 return $value;
             }
         } elseif (is_array($value)) {
             return $this->decryptArray($value);
-        } elseif (is_object($value) && get_class($value) == \stdClass::class) {
+        } elseif (is_object($value) && get_class($value) == stdClass::class) {
             return $this->decryptObject($value);
         } else {
             throw new ApplicationException(
@@ -304,28 +229,16 @@ class ObjectEncryptor
         }
     }
 
-    /**
-     * @param \stdClass $data
-     * @return \stdClass
-     * @throws ApplicationException
-     * @throws UserException
-     */
-    private function decryptObject(\stdClass $data)
+    private function decryptObject(stdClass $data): stdClass
     {
-        $result = new \stdClass();
+        $result = new stdClass();
         foreach (get_object_vars($data) as $key => $value) {
             $result->{$key} = $this->decryptItem($key, $value);
         }
         return $result;
     }
 
-    /**
-     * @param array $data
-     * @return array
-     * @throws ApplicationException
-     * @throws UserException
-     */
-    private function decryptArray(array $data)
+    private function decryptArray(array $data): array
     {
         $result = [];
         foreach ($data as $key => $value) {
