@@ -3,10 +3,6 @@
 namespace Keboola\ObjectEncryptor;
 
 use Keboola\ObjectEncryptor\Exception\ApplicationException;
-use Keboola\ObjectEncryptor\Legacy\Encryptor;
-use Keboola\ObjectEncryptor\Legacy\Wrapper\BaseWrapper;
-use Keboola\ObjectEncryptor\Legacy\Wrapper\ComponentProjectWrapper;
-use Keboola\ObjectEncryptor\Legacy\Wrapper\ComponentWrapper as LegacyComponentWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ComponentAKVWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ComponentKMSWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ConfigurationAKVWrapper;
@@ -18,66 +14,23 @@ use Keboola\ObjectEncryptor\Wrapper\GenericKMSWrapper;
 
 class ObjectEncryptorFactory
 {
-    /**
-     * @var null|string Encryption key for KBC::Encrypted ciphers.
-     */
-    private $keyVersion1 = null;
+    private ?string $stackId = null;
+    private ?string $projectId = null;
+    private ?string $componentId = null;
+    private ?string $configurationId = null;
+    private ?string $kmsKeyId;
+    private ?string $kmsKeyRegion;
+    private ?string $akvUrl;
 
     /**
-     * @var null|string Encryption key for legacy ciphers.
+     * @param ?string $kmsKeyId KMS Key ID Encryption key for KBC::Secure ciphers.
+     * @param ?string $kmsRegion KMS Key Region.
+     * @param ?string $akvUrl Azure Key Vault URL.
      */
-    private $keyVersion0 = null;
-
-    /**
-     * @var null|string Id of KBC Stack.
-     */
-    private $stackId = null;
-
-    /**
-     * @var null|string Id of KBC Project.
-     */
-    private $projectId = null;
-
-    /**
-     * @var null|string Id of current component.
-     */
-    private $componentId = null;
-
-    /**
-     * @var null|string Id of current configuration.
-     */
-    private $configurationId = null;
-
-    /**
-     * @var string|null AWS KMS Key id (ARN) or alias (prefix with 'alias/').
-     */
-    private $kmsKeyId = null;
-
-    /**
-     * @var string|null AWS KMS region.
-     */
-    private $kmsKeyRegion = null;
-
-    /**
-     * @var string|null AKV Vault URL.
-     */
-    private $akvUrl = null;
-
-    /**
-     * ObjectEncryptorFactory constructor.
-     * @param string $kmsKeyId KMS Key ID Encryption key for KBC::Secure ciphers.
-     * @param string $kmsRegion KMS Key Region.
-     * @param string $keyVersion1 Encryption key for KBC::Encrypted ciphers.
-     * @param string $keyVersion0 Encryption key for legacy ciphers.
-     * @param string $akvUrl Azure Key Vault URL.
-     */
-    public function __construct($kmsKeyId, $kmsRegion, $keyVersion1, $keyVersion0, $akvUrl)
+    public function __construct(?string $kmsKeyId, ?string $kmsRegion, ?string $akvUrl)
     {
-        // No logic here, this constructor is exception-less so as not to leak keys in stack trace
         $this->kmsKeyId = $kmsKeyId;
         $this->kmsKeyRegion = $kmsRegion;
-        $this->keyVersion1 = $keyVersion1;
-        $this->keyVersion0 = $keyVersion0;
         $this->akvUrl = $akvUrl;
     }
 
@@ -85,7 +38,7 @@ class ObjectEncryptorFactory
      * @param string|null $componentId Id of current component.
      * @throws ApplicationException
      */
-    public function setComponentId($componentId)
+    public function setComponentId(?string $componentId): void
     {
         if (!is_null($componentId) && !is_scalar($componentId)) {
             throw new ApplicationException('Invalid component id.');
@@ -97,7 +50,7 @@ class ObjectEncryptorFactory
      * @param string|null $configurationId Id of current configuration.
      * @throws ApplicationException
      */
-    public function setConfigurationId($configurationId)
+    public function setConfigurationId(?string $configurationId): void
     {
         if (!is_null($configurationId) && !is_scalar($configurationId)) {
             throw new ApplicationException('Invalid configuration id.');
@@ -109,7 +62,7 @@ class ObjectEncryptorFactory
      * @param string|null $projectId Id of KBC Project.
      * @throws ApplicationException
      */
-    public function setProjectId($projectId)
+    public function setProjectId(?string $projectId): void
     {
         if (!is_null($projectId) && !is_scalar($projectId)) {
             throw new ApplicationException('Invalid project id.');
@@ -121,7 +74,7 @@ class ObjectEncryptorFactory
      * @param string|null $stackId Id of KBC Stack.
      * @throws ApplicationException
      */
-    public function setStackId($stackId)
+    public function setStackId(?string $stackId): void
     {
         if (!is_null($stackId) && !is_scalar($stackId)) {
             throw new ApplicationException('Invalid stack id.');
@@ -132,19 +85,8 @@ class ObjectEncryptorFactory
     /**
      * @throws ApplicationException
      */
-    private function validateState()
+    private function validateState(): void
     {
-        if (!is_null($this->keyVersion0) && !is_string($this->keyVersion0)) {
-            throw new ApplicationException('Invalid key version 0.');
-        }
-        $this->keyVersion0 = substr($this->keyVersion0, 0, 32);
-        if (!$this->keyVersion0) {
-            // For php 5.6 compatibility
-            $this->keyVersion0 = '';
-        }
-        if (!is_null($this->keyVersion1) && !is_string($this->keyVersion1)) {
-            throw new ApplicationException('Invalid key version 1.');
-        }
         if (!is_null($this->kmsKeyId) && !is_string($this->kmsKeyId)) {
             throw new ApplicationException('Invalid KMS key Id.');
         }
@@ -160,31 +102,7 @@ class ObjectEncryptorFactory
      * @param ObjectEncryptor $encryptor
      * @throws ApplicationException
      */
-    private function addLegacyWrappers($encryptor)
-    {
-        $wrapper = new BaseWrapper();
-        $wrapper->setKey((string)$this->keyVersion1);
-        $encryptor->pushWrapper($wrapper);
-        if ($this->componentId !== null) {
-            $wrapper = new LegacyComponentWrapper();
-            $wrapper->setKey((string)$this->keyVersion1);
-            $wrapper->setComponentId($this->componentId);
-            $encryptor->pushWrapper($wrapper);
-            if ($this->projectId !== null) {
-                $wrapper = new ComponentProjectWrapper();
-                $wrapper->setKey((string)$this->keyVersion1);
-                $wrapper->setComponentId($this->componentId);
-                $wrapper->setProjectId($this->projectId);
-                $encryptor->pushWrapper($wrapper);
-            }
-        }
-    }
-
-    /**
-     * @param ObjectEncryptor $encryptor
-     * @throws ApplicationException
-     */
-    private function addKMSWrappers($encryptor)
+    private function addKMSWrappers(ObjectEncryptor $encryptor): void
     {
         $wrapper = new GenericKMSWrapper();
         $wrapper->setKMSKeyId((string)$this->kmsKeyId);
@@ -224,7 +142,7 @@ class ObjectEncryptorFactory
      * @param ObjectEncryptor $encryptor
      * @throws ApplicationException
      */
-    private function addAKVWrappers($encryptor)
+    private function addAKVWrappers(ObjectEncryptor $encryptor): void
     {
         $wrapper = new GenericAKVWrapper();
         $wrapper->setKeyVaultUrl((string) $this->akvUrl);
@@ -256,25 +174,11 @@ class ObjectEncryptorFactory
         }
     }
 
-    /**
-     * @param bool $createLegacyEncryptorIfAvailable
-     * @return ObjectEncryptor Object encryptor instance.
-     * @throws ApplicationException
-     */
-    public function getEncryptor($createLegacyEncryptorIfAvailable = false)
+    public function getEncryptor(): ObjectEncryptor
     {
         $this->validateState();
-        if ($this->keyVersion0 && function_exists('mcrypt_module_open') && $createLegacyEncryptorIfAvailable) {
-            $legacyEncryptor = new Encryptor($this->keyVersion0);
-        } else {
-            $legacyEncryptor = null;
-        }
 
-        $encryptor = new ObjectEncryptor($legacyEncryptor);
-        if ($this->keyVersion1) {
-            $this->addLegacyWrappers($encryptor);
-        }
-
+        $encryptor = new ObjectEncryptor();
         if ($this->akvUrl) {
             $this->addAKVWrappers($encryptor);
         }
