@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Keboola\ObjectEncryptor\Wrapper;
 
 use Keboola\ObjectEncryptor\Exception\ApplicationException;
@@ -46,7 +48,7 @@ class ComponentAKVWrapper extends GenericAKVWrapper
     /**
      * @inheritdoc
      */
-    public function getPrefix()
+    public function getPrefix(): string
     {
         return 'KBC::ComponentSecureKV::';
     }
